requirements.txt
requests
pybase64
Pillow
python-dotenv
openai
#simpleaudio
keyboard
pygame 
